Move page loading out of tests
