+++ package.json
@@ -41,8 +41,8 @@
     "@rollup/pluginutils": "4.0.0",
     "@types/chrome": "0.0.124",
     "@types/jest": "26.0.14",
-    "@types/puppeteer-core": "2.0.0",
     "@types/node": "14.11.2",
+    "@types/puppeteer-core": "2.0.0",
     "@typescript-eslint/eslint-plugin": "4.3.0",
     "@typescript-eslint/parser": "4.3.0",
     "chokidar": "3.4.2",
