+++ tests/browser/globals.d.ts
@@ -1,0 +1,6 @@
+import {Page} from 'puppeteer-core';
+
+declare global {
+    const loadTestPage: (paths: {[path: string]: string}) => Promise<void>;
+    const page: Page;
+}
